requirements.txt
streamlit>=1.28.0
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pyyaml>=6.0.0
